chore: sorting of methods

+++ app/API/users.py
@@ -14,6 +14,32 @@ pwd_hasher = CryptContext(schemes=["bcrypt"], deprecated="auto")
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/login")
 router = APIRouter()
 
+@router.post(
+        "", 
+        status_code = status.HTTP_201_CREATED,
+        response_model = User
+    )
+async def register(
+        user: UserCredentials
+    ):
+    new_user = {
+        "username": user.username,
+        "password": pwd_hasher.hash(user.password),
+        "favourites": []
+    }
+
+    user = user_collection.find_one({"username": new_user["username"]})
+
+    if user:
+        raise HTTPException(
+                status_code=status.HTTP_409_CONFLICT, 
+                detail="Specified username already exists"
+            )
+
+    _id = user_collection.insert_one(new_user)
+    new_user["id"] = _id.inserted_id
+    return new_user
+
 async def get_current_user(
         jwt_token: str = Depends(oauth2_scheme)
     ):
@@ -52,26 +78,6 @@ async def get_current_user_profile(
     return current_user
 
 @router.get(
-        "/{username}", 
-        response_model = UserInDB, 
-        status_code = status.HTTP_200_OK
-    )
-async def get_user_by_username(
-        username: str, 
-        current_user: User = Depends(get_current_user)
-    ):
-    if(current_user["username"] == username):
-        return current_user
-    else:
-        
-        raise HTTPException(
-                status_code=status.HTTP_403_FORBIDDEN, 
-                detail="You don not have permission to use this resource",
-                )#TODO VER QUE HACER CON ESTE METODO SI VUELA O LO DEJAMOS, LO MISMO CON GET A /USERS
-                #TODO CON PAGINACION EL GET /USERS
-
-
-@router.get(
         "/me/documents", 
         response_model = PaginatedDocumentFav, 
         status_code = status.HTTP_200_OK
@@ -108,50 +114,16 @@ async def get_current_user_favourites(
         documents = [str(id) for id in get_documents]
     ) #TODO devolver los documentos
 
-     
-
-@router.post(
-        "", 
-        status_code = status.HTTP_201_CREATED,
-        response_model = User
-    )
-async def register(
-        user: UserCredentials
-    ):
-    new_user = {
-        "username": user.username,
-        "password": pwd_hasher.hash(user.password),
-        "favourites": []
-    }
-
-    user = user_collection.find_one({"username": new_user["username"]})
-
-    if user:
-        raise HTTPException(
-                status_code=status.HTTP_409_CONFLICT, 
-                detail="Specified username already exists"
-            )
-
-    _id = user_collection.insert_one(new_user)
-    new_user["id"] = _id.inserted_id
-    return new_user
-	    
 
 @router.delete(
-        "/{id}", 
+        "", 
         status_code = status.HTTP_200_OK
     )
 async def delete_user_by_username(
-        username: str, 
         current_user: User = Depends(get_current_user),
     ):
-    if not current_user["username"] == username:
-        raise HTTPException(
-                status_code=status.HTTP_403_FORBIDDEN, 
-                detail="You don not have permission to use this resource"
-            )
-    documents = List(document_collection.find({"author.username": username}))
+    documents = document_collection.find({"author": current_user["username"]})
     for document in documents:
         document_collection.delete_one({"_id": document["_id"]})
-    user_collection.delete_one({"username": username})
+    user_collection.delete_one({"username": current_user["username"]})
     return {}
